[modify] 新規登録画面とログイン画面のレイアウト修正
