+++ lib/ui/auth/view/email_login_page.dart
@@ -1,4 +1,9 @@
 import 'package:animeishi/ui/home/view/home_page.dart';
+import 'package:animeishi/ui/auth/components/email_login_validation.dart';
+import 'package:animeishi/ui/auth/components/email_login_dialogs.dart';
+import 'package:animeishi/ui/auth/components/email_login_animations.dart';
+import 'package:animeishi/ui/auth/components/email_login_widgets.dart';
+import 'package:animeishi/ui/auth/components/auth_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'dart:async';
@@ -81,28 +86,12 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
     super.dispose();
   }
 
-  String? _validateEmail() {
-    if (emailController.text.isEmpty) {
-      return 'メールアドレスを入力してください';
-    } else if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(emailController.text)) {
-      return '正しいメールアドレスを入力してください';
-    }
-    return null;
-  }
-
-  String? _validatePassword() {
-    if (passwordController.text.isEmpty) {
-      return 'パスワードを入力してください';
-    }
-    return null;
-  }
-
   Future<void> _login() async {
-    final emailError = _validateEmail();
-    final passwordError = _validatePassword();
+    final emailError = EmailLoginValidation.validateEmail(emailController.text);
+    final passwordError = EmailLoginValidation.validatePassword(passwordController.text);
     
     if (emailError != null || passwordError != null) {
-      _showValidationDialog(emailError, passwordError);
+      EmailLoginDialogs.showValidationDialog(context, emailError, passwordError);
       return;
     }
 
@@ -120,7 +109,7 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
       if (user != null) {
         print("ログインしました ${user.email}, ${user.uid}");
         
-        _showSuccessMessage();
+        EmailLoginDialogs.showSuccessMessage(context);
         
         await Future.delayed(Duration(milliseconds: 1500));
         
@@ -135,10 +124,10 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
       print('ログインエラー: $e');
       
       setState(() {
-        errorMessage = _getErrorMessage(e.toString());
+        errorMessage = EmailLoginValidation.getFirebaseErrorMessage(e.toString());
       });
       
-      _showErrorDialog(errorMessage);
+      EmailLoginDialogs.showErrorDialog(context, errorMessage);
     } finally {
       if (mounted) {
         setState(() {
@@ -155,543 +144,46 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
   }
 
   Future<void> _resetPassword() async {
-    final emailError = _validateEmail();
+    final emailError = EmailLoginValidation.validateEmail(emailController.text);
     
     if (emailError != null) {
-      _showValidationDialog(emailError, null);
+      EmailLoginDialogs.showValidationDialog(context, emailError, null);
       return;
     }
 
     try {
       await FirebaseAuth.instance.sendPasswordResetEmail(email: emailController.text.trim());
       
-      _showInfoDialog(
+      EmailLoginDialogs.showInfoDialog(
+        context,
         'パスワードリセット',
         '${emailController.text}へパスワードリセット用のメールを送信しました',
         Icons.email_outlined,
         Colors.blue,
       );
     } catch (e) {
-      _showErrorDialog('パスワードリセットメールの送信に失敗しました');
+      EmailLoginDialogs.showErrorDialog(context, 'パスワードリセットメールの送信に失敗しました');
     }
-  }
-
-  void _showValidationDialog(String? emailError, String? passwordError) {
-    showDialog(
-      context: context,
-      barrierDismissible: true,
-      builder: (BuildContext context) {
-        return Dialog(
-          backgroundColor: Colors.transparent,
-          child: Container(
-            padding: EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: [
-                  Colors.white.withOpacity(0.95),
-                  Colors.white.withOpacity(0.9),
-                ],
-              ),
-              borderRadius: BorderRadius.circular(24),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.1),
-                  blurRadius: 20,
-                  offset: Offset(0, 10),
-                ),
-              ],
-            ),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Container(
-                  width: 80,
-                  height: 80,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        Colors.orange.shade300,
-                        Colors.orange.shade500,
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(40),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.orange.withOpacity(0.3),
-                        blurRadius: 15,
-                        offset: Offset(0, 5),
-                      ),
-                    ],
-                  ),
-                  child: Icon(
-                    Icons.edit_note_rounded,
-                    color: Colors.white,
-                    size: 40,
-                  ),
-                ),
-                
-                SizedBox(height: 20),
-                
-                Text(
-                  '入力内容をご確認ください',
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w700,
-                    color: Color(0xFF2D3748),
-                  ),
-                ),
-                
-                SizedBox(height: 16),
-                
-                if (emailError != null || passwordError != null) ...[
-                  Container(
-                    padding: EdgeInsets.all(16),
-                    decoration: BoxDecoration(
-                      color: Colors.orange.shade50,
-                      borderRadius: BorderRadius.circular(12),
-                      border: Border.all(
-                        color: Colors.orange.withOpacity(0.3),
-                        width: 1,
-                      ),
-                    ),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        if (emailError != null) ...[
-                          Row(
-                            children: [
-                              Icon(
-                                Icons.email_outlined,
-                                color: Colors.orange.shade600,
-                                size: 18,
-                              ),
-                              SizedBox(width: 8),
-                              Expanded(
-                                child: Text(
-                                  emailError,
-                                  style: TextStyle(
-                                    color: Colors.orange.shade700,
-                                    fontSize: 14,
-                                    fontWeight: FontWeight.w500,
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                          if (passwordError != null) SizedBox(height: 12),
-                        ],
-                        if (passwordError != null) ...[
-                          Row(
-                            children: [
-                              Icon(
-                                Icons.lock_outline,
-                                color: Colors.orange.shade600,
-                                size: 18,
-                              ),
-                              SizedBox(width: 8),
-                              Expanded(
-                                child: Text(
-                                  passwordError,
-                                  style: TextStyle(
-                                    color: Colors.orange.shade700,
-                                    fontSize: 14,
-                                    fontWeight: FontWeight.w500,
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ],
-                    ),
-                  ),
-                ],
-                
-                SizedBox(height: 24),
-                
-                Container(
-                  width: double.infinity,
-                  height: 48,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        Colors.orange.shade400,
-                        Colors.orange.shade600,
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(24),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.orange.withOpacity(0.3),
-                        blurRadius: 10,
-                        offset: Offset(0, 4),
-                      ),
-                    ],
-                  ),
-                  child: Material(
-                    color: Colors.transparent,
-                    child: InkWell(
-                      borderRadius: BorderRadius.circular(24),
-                      onTap: () => Navigator.of(context).pop(),
-                      child: Center(
-                        child: Text(
-                          '確認しました',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  void _showSuccessMessage() {
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Row(
-          children: [
-            Icon(Icons.check_circle_outline, color: Colors.white, size: 20),
-            SizedBox(width: 8),
-            Text('ログインしました！'),
-          ],
-        ),
-        backgroundColor: Colors.green.shade600,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
-        margin: EdgeInsets.all(16),
-        duration: Duration(seconds: 2),
-      ),
-    );
-  }
-
-  void _showErrorDialog(String error) {
-    showDialog(
-      context: context,
-      barrierDismissible: true,
-      builder: (BuildContext context) {
-        return Dialog(
-          backgroundColor: Colors.transparent,
-          child: Container(
-            padding: EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: [
-                  Colors.white.withOpacity(0.95),
-                  Colors.white.withOpacity(0.9),
-                ],
-              ),
-              borderRadius: BorderRadius.circular(24),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.1),
-                  blurRadius: 20,
-                  offset: Offset(0, 10),
-                ),
-              ],
-            ),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Container(
-                  width: 80,
-                  height: 80,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        Colors.red.shade300,
-                        Colors.red.shade500,
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(40),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.red.withOpacity(0.3),
-                        blurRadius: 15,
-                        offset: Offset(0, 5),
-                      ),
-                    ],
-                  ),
-                  child: Icon(
-                    Icons.error_outline,
-                    color: Colors.white,
-                    size: 40,
-                  ),
-                ),
-                
-                SizedBox(height: 20),
-                
-                Text(
-                  'ログインエラー',
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w700,
-                    color: Color(0xFF2D3748),
-                  ),
-                ),
-                
-                SizedBox(height: 16),
-                
-                Container(
-                  padding: EdgeInsets.all(16),
-                  decoration: BoxDecoration(
-                    color: Colors.red.shade50,
-                    borderRadius: BorderRadius.circular(12),
-                    border: Border.all(
-                      color: Colors.red.withOpacity(0.3),
-                      width: 1,
-                    ),
-                  ),
-                  child: Text(
-                    error,
-                    style: TextStyle(
-                      color: Colors.red.shade700,
-                      fontSize: 14,
-                      fontWeight: FontWeight.w500,
-                    ),
-                    textAlign: TextAlign.center,
-                  ),
-                ),
-                
-                SizedBox(height: 24),
-                
-                Container(
-                  width: double.infinity,
-                  height: 48,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        Colors.red.shade400,
-                        Colors.red.shade600,
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(24),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.red.withOpacity(0.3),
-                        blurRadius: 10,
-                        offset: Offset(0, 4),
-                      ),
-                    ],
-                  ),
-                  child: Material(
-                    color: Colors.transparent,
-                    child: InkWell(
-                      borderRadius: BorderRadius.circular(24),
-                      onTap: () => Navigator.of(context).pop(),
-                      child: Center(
-                        child: Text(
-                          '確認しました',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  void _showInfoDialog(String title, String message, IconData icon, Color color) {
-    showDialog(
-      context: context,
-      barrierDismissible: true,
-      builder: (BuildContext context) {
-        return Dialog(
-          backgroundColor: Colors.transparent,
-          child: Container(
-            padding: EdgeInsets.all(24),
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: [
-                  Colors.white.withOpacity(0.95),
-                  Colors.white.withOpacity(0.9),
-                ],
-              ),
-              borderRadius: BorderRadius.circular(24),
-              boxShadow: [
-                BoxShadow(
-                  color: Colors.black.withOpacity(0.1),
-                  blurRadius: 20,
-                  offset: Offset(0, 10),
-                ),
-              ],
-            ),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Container(
-                  width: 80,
-                  height: 80,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        color.withOpacity(0.7),
-                        color.withOpacity(0.9),
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(40),
-                    boxShadow: [
-                      BoxShadow(
-                        color: color.withOpacity(0.3),
-                        blurRadius: 15,
-                        offset: Offset(0, 5),
-                      ),
-                    ],
-                  ),
-                  child: Icon(
-                    icon,
-                    color: Colors.white,
-                    size: 40,
-                  ),
-                ),
-                
-                SizedBox(height: 20),
-                
-                Text(
-                  title,
-                  style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w700,
-                    color: Color(0xFF2D3748),
-                  ),
-                ),
-                
-                SizedBox(height: 16),
-                
-                Container(
-                  padding: EdgeInsets.all(16),
-                  decoration: BoxDecoration(
-                    color: color.withOpacity(0.1),
-                    borderRadius: BorderRadius.circular(12),
-                    border: Border.all(
-                      color: color.withOpacity(0.3),
-                      width: 1,
-                    ),
-                  ),
-                  child: Text(
-                    message,
-                    style: TextStyle(
-                      color: color.withOpacity(0.8),
-                      fontSize: 14,
-                      fontWeight: FontWeight.w500,
-                    ),
-                    textAlign: TextAlign.center,
-                  ),
-                ),
-                
-                SizedBox(height: 24),
-                
-                Container(
-                  width: double.infinity,
-                  height: 48,
-                  decoration: BoxDecoration(
-                    gradient: LinearGradient(
-                      colors: [
-                        color.withOpacity(0.8),
-                        color.withOpacity(1.0),
-                      ],
-                    ),
-                    borderRadius: BorderRadius.circular(24),
-                    boxShadow: [
-                      BoxShadow(
-                        color: color.withOpacity(0.3),
-                        blurRadius: 10,
-                        offset: Offset(0, 4),
-                      ),
-                    ],
-                  ),
-                  child: Material(
-                    color: Colors.transparent,
-                    child: InkWell(
-                      borderRadius: BorderRadius.circular(24),
-                      onTap: () => Navigator.of(context).pop(),
-                      child: Center(
-                        child: Text(
-                          '確認しました',
-                          style: TextStyle(
-                            color: Colors.white,
-                            fontSize: 16,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  String _getErrorMessage(String error) {
-    if (error.contains('user-not-found')) {
-      return 'このメールアドレスは登録されていません';
-    } else if (error.contains('wrong-password')) {
-      return 'パスワードが間違っています';
-    } else if (error.contains('invalid-email')) {
-      return 'メールアドレスの形式が正しくありません';
-    } else if (error.contains('user-disabled')) {
-      return 'このアカウントは無効化されています';
-    } else if (error.contains('too-many-requests')) {
-      return 'ログイン試行回数が多すぎます。しばらく待ってからお試しください';
-    }
-    return 'ログインに失敗しました。メールアドレスとパスワードをご確認ください';
   }
 
   @override
   Widget build(BuildContext context) {
     final size = MediaQuery.of(context).size;
     final isSmallScreen = size.height < 700;
-    
+
     return Scaffold(
       body: Stack(
         children: [
-          Container(
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                begin: Alignment.topLeft,
-                end: Alignment.bottomRight,
-                colors: [
-                  Color(0xFFD6BCFA),
-                  Color(0xFFBFDBFE),
-                  Color(0xFFFBCFE8),
-                  Color(0xFFD1FAE5),
-                ],
-              ),
-            ),
-          ),
+          // 背景グラデーション
+          AuthWidgets.buildBackgroundGradient(),
           
+          // パーティクルアニメーション
           CustomPaint(
             size: size,
             painter: AnimationPainter(particles),
           ),
           
+          // メインコンテンツ
           SafeArea(
             child: SingleChildScrollView(
               physics: BouncingScrollPhysics(),
@@ -700,98 +192,42 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
                   minHeight: size.height - MediaQuery.of(context).padding.top - MediaQuery.of(context).padding.bottom,
                 ),
                 child: Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 32.0, vertical: 20.0),
+                  padding: const EdgeInsets.symmetric(horizontal: 32.0, vertical: 40.0),
                   child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
                     children: [
-                      Row(
-                        children: [
-                          IconButton(
-                            onPressed: () => Navigator.pop(context),
-                            icon: Icon(
-                              Icons.arrow_back_ios,
-                              color: Color(0xFF2D3748),
-                              size: 24,
-                            ),
-                          ),
-                        ],
+                      // 戻るボタン
+                      AuthWidgets.buildBackButton(
+                        onPressed: () => Navigator.of(context).pop(),
                       ),
                       
-                      SizedBox(height: isSmallScreen ? 20 : 40),
-                      
+                      // タイトル部分
                       FadeTransition(
                         opacity: _fadeAnimation,
                         child: SlideTransition(
                           position: _slideAnimation,
-                          child: Column(
-                            children: [
-                              Container(
-                                width: isSmallScreen ? 80 : 100,
-                                height: isSmallScreen ? 80 : 100,
-                                decoration: BoxDecoration(
-                                  gradient: LinearGradient(
-                                    begin: Alignment.topLeft,
-                                    end: Alignment.bottomRight,
-                                    colors: [
-                                      Color(0xFF667eea).withOpacity(0.8),
-                                      Color(0xFF764ba2).withOpacity(0.9),
-                                    ],
-                                  ),
-                                  borderRadius: BorderRadius.circular(20),
-                                  boxShadow: [
-                                    BoxShadow(
-                                      color: Color(0xFF667eea).withOpacity(0.25),
-                                      blurRadius: 20,
-                                      offset: Offset(0, 8),
-                                    ),
-                                  ],
-                                ),
-                                child: Icon(
-                                  Icons.login,
-                                  size: isSmallScreen ? 35 : 45,
-                                  color: Colors.white,
-                                ),
-                              ),
-                              SizedBox(height: isSmallScreen ? 24 : 32),
-                              
-                              Text(
-                                'ログイン',
-                                style: TextStyle(
-                                  fontSize: isSmallScreen ? 32 : 40,
-                                  fontWeight: FontWeight.w800,
-                                  color: Color(0xFF2D3748),
-                                  letterSpacing: 1.0,
-                                  shadows: [
-                                    Shadow(
-                                      color: Colors.black.withOpacity(0.1),
-                                      offset: Offset(0, 2),
-                                      blurRadius: 4,
-                                    ),
-                                  ],
-                                ),
-                              ),
-                              SizedBox(height: 8),
-                              
-                              Text(
-                                'アカウントにログインしてアニ名刺を楽しみましょう',
-                                style: TextStyle(
-                                  fontSize: isSmallScreen ? 14 : 16,
-                                  color: Color(0xFF718096),
-                                  fontWeight: FontWeight.w500,
-                                ),
-                                textAlign: TextAlign.center,
-                              ),
+                          child: AuthWidgets.buildPageTitle(
+                            title: 'ログイン',
+                            subtitle: 'アカウントにサインインしてください',
+                            icon: Icons.login_rounded,
+                            iconColors: [
+                              Color(0xFF667eea).withOpacity(0.8),
+                              Color(0xFF764ba2).withOpacity(0.9),
                             ],
+                            isSmallScreen: isSmallScreen,
                           ),
                         ),
                       ),
                       
                       SizedBox(height: isSmallScreen ? 40 : 60),
                       
+                      // フォーム部分
                       FadeTransition(
                         opacity: _fadeAnimation,
                         child: Column(
                           children: [
-                            _buildModernTextField(
+                            // メールアドレス入力
+                            EmailLoginWidgets.buildInputField(
                               controller: emailController,
                               label: 'メールアドレス',
                               icon: Icons.email_outlined,
@@ -800,7 +236,8 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
                             
                             SizedBox(height: 20),
                             
-                            _buildModernTextField(
+                            // パスワード入力
+                            EmailLoginWidgets.buildInputField(
                               controller: passwordController,
                               label: 'パスワード',
                               icon: Icons.lock_outline,
@@ -820,35 +257,40 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
                             
                             SizedBox(height: 32),
                             
-                            _buildLoginButton(),
+                            // ログインボタン
+                            EmailLoginWidgets.buildLoginButton(
+                              isLoading: isLoading,
+                              onPressed: _login,
+                            ),
                             
-                            SizedBox(height: 16),
+                            SizedBox(height: 20),
                             
+                            // セカンダリボタン
                             Row(
                               children: [
                                 Expanded(
-                                  child: _buildSecondaryButton(
+                                  child: EmailLoginWidgets.buildSecondaryButton(
                                     text: 'テストログイン',
-                                    icon: Icons.science_outlined,
+                                    icon: Icons.bug_report,
                                     onPressed: _testLogin,
                                     colors: [
-                                      Colors.blue.shade400,
-                                      Colors.blue.shade600,
+                                      Color(0xFF38b2ac).withOpacity(0.8),
+                                      Color(0xFF319795).withOpacity(0.9),
                                     ],
+                                    isLoading: isLoading,
                                   ),
                                 ),
-                                
                                 SizedBox(width: 12),
-                                
                                 Expanded(
-                                  child: _buildSecondaryButton(
+                                  child: EmailLoginWidgets.buildSecondaryButton(
                                     text: 'パスワードリセット',
-                                    icon: Icons.refresh_outlined,
+                                    icon: Icons.refresh,
                                     onPressed: _resetPassword,
                                     colors: [
-                                      Colors.purple.shade400,
-                                      Colors.purple.shade600,
+                                      Color(0xFFf093fb).withOpacity(0.8),
+                                      Color(0xFFf5576c).withOpacity(0.9),
                                     ],
+                                    isLoading: isLoading,
                                   ),
                                 ),
                               ],
@@ -865,205 +307,6 @@ class _EmailLoginPageState extends State<EmailLoginPage> with TickerProviderStat
             ),
           ),
         ],
-      ),
-    );
-  }
-
-  Widget _buildModernTextField({
-    required TextEditingController controller,
-    required String label,
-    required IconData icon,
-    TextInputType? keyboardType,
-    bool obscureText = false,
-    Widget? suffixIcon,
-  }) {
-    return Container(
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-          colors: [
-            Colors.white.withOpacity(0.95),
-            Colors.white.withOpacity(0.85),
-          ],
-        ),
-        borderRadius: BorderRadius.circular(20),
-        border: Border.all(
-          color: Colors.white.withOpacity(0.3),
-          width: 1.5,
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.08),
-            blurRadius: 15,
-            offset: Offset(0, 5),
-          ),
-        ],
-      ),
-      child: TextFormField(
-        controller: controller,
-        keyboardType: keyboardType,
-        obscureText: obscureText,
-        style: TextStyle(
-          fontSize: 16,
-          color: Color(0xFF2D3748),
-          fontWeight: FontWeight.w500,
-        ),
-        decoration: InputDecoration(
-          labelText: label,
-          labelStyle: TextStyle(
-            color: Color(0xFF718096),
-            fontSize: 15,
-            fontWeight: FontWeight.w400,
-          ),
-          floatingLabelStyle: TextStyle(
-            color: Color(0xFF667eea),
-            fontSize: 16,
-            fontWeight: FontWeight.w600,
-          ),
-          prefixIcon: Container(
-            margin: EdgeInsets.only(left: 12, right: 8),
-            child: Icon(
-              icon,
-              color: Color(0xFF718096),
-              size: 22,
-            ),
-          ),
-          suffixIcon: suffixIcon,
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(20),
-            borderSide: BorderSide.none,
-          ),
-          focusedBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(20),
-            borderSide: BorderSide(
-              color: Color(0xFF667eea).withOpacity(0.8),
-              width: 2,
-            ),
-          ),
-          filled: true,
-          fillColor: Colors.transparent,
-          contentPadding: EdgeInsets.symmetric(horizontal: 24, vertical: 20),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildLoginButton() {
-    return Container(
-      width: double.infinity,
-      height: 56,
-      decoration: BoxDecoration(
-        gradient: LinearGradient(
-          colors: [
-            Color(0xFF667eea).withOpacity(0.8),
-            Color(0xFF764ba2).withOpacity(0.9),
-          ],
-        ),
-        borderRadius: BorderRadius.circular(16),
-        boxShadow: [
-          BoxShadow(
-            color: Color(0xFF667eea).withOpacity(0.3),
-            blurRadius: 16,
-            offset: Offset(0, 6),
-          ),
-        ],
-      ),
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          borderRadius: BorderRadius.circular(16),
-          onTap: isLoading ? null : _login,
-          child: Container(
-            padding: EdgeInsets.symmetric(horizontal: 24),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                if (isLoading) ...[
-                  SizedBox(
-                    width: 20,
-                    height: 20,
-                    child: CircularProgressIndicator(
-                      strokeWidth: 2,
-                      valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-                    ),
-                  ),
-                  SizedBox(width: 12),
-                ] else ...[
-                  Icon(
-                    Icons.login,
-                    color: Colors.white,
-                    size: 24,
-                  ),
-                  SizedBox(width: 12),
-                ],
-                Text(
-                  isLoading ? 'ログイン中...' : 'ログイン',
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 18,
-                    fontWeight: FontWeight.w600,
-                    letterSpacing: 0.5,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSecondaryButton({
-    required String text,
-    required IconData icon,
-    required VoidCallback onPressed,
-    required List<Color> colors,
-  }) {
-    return Container(
-      height: 48,
-      decoration: BoxDecoration(
-        gradient: LinearGradient(colors: colors),
-        borderRadius: BorderRadius.circular(14),
-        boxShadow: [
-          BoxShadow(
-            color: colors[0].withOpacity(0.3),
-            blurRadius: 12,
-            offset: Offset(0, 4),
-          ),
-        ],
-      ),
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          borderRadius: BorderRadius.circular(14),
-          onTap: isLoading ? null : onPressed,
-          child: Container(
-            padding: EdgeInsets.symmetric(horizontal: 16),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Icon(
-                  icon,
-                  color: Colors.white,
-                  size: 18,
-                ),
-                SizedBox(width: 8),
-                Flexible(
-                  child: Text(
-                    text,
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontSize: 14,
-                      fontWeight: FontWeight.w600,
-                    ),
-                    overflow: TextOverflow.ellipsis,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
       ),
     );
   }
